+++ matrix-sdk/src/main/java/org/matrix/androidsdk/rest/client/AccountDataRestClient.java
@@ -23,9 +23,6 @@ import org.matrix.androidsdk.core.callback.ApiCallback;
 import org.matrix.androidsdk.rest.api.AccountDataApi;
 import org.matrix.androidsdk.rest.callback.RestAdapterCallback;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class AccountDataRestClient extends RestClient<AccountDataApi> {
     /**
      * {@inheritDoc}
@@ -56,24 +53,4 @@ public class AccountDataRestClient extends RestClient<AccountDataApi> {
                 }));
     }
 
-    /**
-     * Gets a bearer token from the homeserver that the user can
-     * present to a third party in order to prove their ownership
-     * of the Matrix account they are logged into.
-     *
-     * @param userId   the user id
-     * @param callback the asynchronous callback called when finished
-     */
-    public void openIdToken(final String userId, final ApiCallback<Map<Object, Object>> callback) {
-        final String description = "openIdToken userId : " + userId;
-
-        mApi.openIdToken(userId, new HashMap<>())
-                .enqueue(new RestAdapterCallback<Map<Object, Object>>(description, mUnsentEventsManager, callback,
-                        new RestAdapterCallback.RequestRetryCallBack() {
-                            @Override
-                            public void onRetry() {
-                                openIdToken(userId, callback);
-                            }
-                        }));
-    }
 }