+++ CHANGES.rst
@@ -12,7 +12,7 @@ Bugfix:
  -
 
 API Change:
- -
+ - `MXSession.openIdToken()` callback has a more typed parameter
 
 Translations:
  -